refactor: pull data up from Grid.tsx
