+++ src/components/pickup/Grid.tsx
@@ -2,24 +2,9 @@ import React, { useContext, useState } from "react";
 import { DataGrid, GridRowParams, MuiEvent, GridCallbackDetails } from "@mui/x-data-grid";
 import { stepperContext } from "../../providers/StepperProvider";
 
-function createData(id: number, no: number, name: string, self: number, afac: number, total: number, method: string, paid: string, pickup: string) {
-    return { id: id, no: no, name: name, self: self, afac: afac, total: total, method: method, paid: paid, pickup: pickup };
+export interface GridOrder {
+    id: number, no: number, name: string, self: number, afac: number, total: number, method: string, paid: string, pickup: string
 }
-
-const data = [
-    createData(1, 1, "Ava Miller", 1, 4, 5, "Credit Card", "Yes", "Ready"),
-    createData(2, 2, "James Cole", 1, 0, 4, "Credit Card", "Yes", "Ready"),
-    createData(3, 3, "Vivian Eggers", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(4, 4, "Ellijiah Sandis", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(5, 5, "Anjali Sharma", 1, 4, 4, "Credit Card", "No", "Ready"),
-    createData(6, 6, "Sarah Smith", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(7, 7, "Noah Davis", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(8, 8, " Mary Brown", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(9, 9, " Miller", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(10, 10, " Miller", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(11, 11, " Miller", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-    createData(12, 12, " Miller", 1, 4, 4, "Credit Card", "Yes", "Ready"),
-];
 
 const columns = [
     { headerName: "No.", field: "no" },
@@ -47,7 +32,7 @@ const columns = [
 ];
 
 
-export default function Grid(props: {onRowClick: (number) => void}) {
+export default function Grid(props: {onRowClick: (number) => void, data: GridOrder[]}) {
 
     function onRowClick(params: GridRowParams, _: MuiEvent<React.MouseEvent>, _2: GridCallbackDetails): void {
         if (activeStep === 6) {
@@ -59,7 +44,7 @@ export default function Grid(props: {onRowClick: (number) => void}) {
     return (
         <div style={{ height: 500, width: "100%", backgroundColor: "white" }}>
             <DataGrid
-                rows={data}
+                rows={props.data}
                 columns={columns}
                 componentsProps={{
                 }}
